requirements.txt
streamlit   
pycryptodome